requirements.txt
Django==3.2.8
djangorestframework==3.14.0
django-cors-headers==3.8.0